chore(backend): lib.rs refactor + fix pre-existing clippy lints
Add src/lib.rs (crate `stablecoin_pay`) exposing the modules + AppState at
the crate root, and slim src/main.rs to `use stablecoin_pay::*`, so both the
API server and the upcoming agent binary can share code.

Also clear pre-existing `clippy -- -D warnings` debt with low-risk changes so
the agent build's standard gate can be green: auto-fixable lints (unused
import, map().flatten(), unnecessary cast), plus #[allow(dead_code)] on two
serde-only structs and #[allow(clippy::too_many_arguments)] on six existing
builder/repo functions. No behavior changes.

+++ backend/src/services/webhook.rs
@@ -161,7 +161,7 @@ impl WebhookService {
                     .await?;
 
                     // If we've exhausted retries, mark as failed
-                    if attempt_num >= MAX_ATTEMPTS as i32 {
+                    if attempt_num >= MAX_ATTEMPTS {
                         WebhookEventRepository::mark_failed(&self.pool, event_id, &error_msg)
                             .await?;
                         error!(
